requirements.txt
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pandas==2.0.0
SQLAlchemy==2.0.15
matplotlib==3.7.1
seaborn==0.12.2